# pkg/client/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: info@green-ecolution.de
    name: Green Ecolution
    url: https://green-ecolution.de
  description: This is the API for the Green Ecolution Management System.
  license:
    name: AGPL
    url: https://raw.githubusercontent.com/green-ecolution/green-ecolution-management/develop/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Green Space Management API
  version: develop
servers:
- url: /
paths:
...
  /v1/sensor:
    get:
      description: Get all sensors
      operationId: get-all-sensors
      parameters:
      - description: Page
        in: query
        name: page
        schema:
          type: integer
      - description: Limit
        in: query
        name: limit
        schema:
          type: integer
      - description: Provider
        in: query
        name: provider
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorList'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get all sensors
      tags:
      - Sensor
  /v1/sensor/{sensor_id}:
    delete:
      description: Delete sensor
      operationId: delete-sensor
      parameters:
      - description: Sensor ID
        in: path
        name: sensor_id
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Delete sensor
      tags:
      - Sensor
    get:
      description: Get sensor by ID
      operationId: get-sensor-by-id
      parameters:
      - description: Sensor ID
        in: path
        name: sensor_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get sensor by ID
      tags:
      - Sensor
...
components:
  schemas:
...
    HTTPError:
      example:
        path: path
        code: 7
        method: method
        error: error
      properties:
        code:
          type: integer
        error:
          type: string
        method:
          type: string
        path:
          type: string
      required:
      - code
      - error
      - method
      - path
      type: object
...
    Pagination:
      example:
        next_page: 9
        total_records: 4
        total_pages: 2
        prev_page: 3
        current_page: 7
      properties:
        current_page:
          type: integer
        next_page:
          type: integer
        prev_page:
          type: integer
        total_pages:
          type: integer
        total_records:
          type: integer
      required:
      - current_page
      - next_page
      - prev_page
      - total_pages
      - total_records
      type: object
...
    Sensor:
      example:
        additional_information:
          key: ""
        latest_data:
          updated_at: updated_at
          temperature: 4.145608029883936
          created_at: created_at
          humidity: 2.027123023002322
          watermarks:
          - centibar: 7
            depth: 1
            resistance: 1
          - centibar: 7
            depth: 1
            resistance: 1
          battery: 3.616076749251911
        updated_at: updated_at
        provider: provider
        latitude: 1.4894159098541704
        created_at: created_at
        id: id
        longitude: 6.84685269835264
        status: online
      properties:
        additional_information:
          additionalProperties: true
          type: object
        created_at:
          type: string
        id:
          type: string
        latest_data:
          $ref: '#/components/schemas/SensorData'
        latitude:
          type: number
        longitude:
          type: number
        provider:
          type: string
        status:
          $ref: '#/components/schemas/SensorStatus'
        updated_at:
          type: string
      required:
      - additional_information
      - created_at
      - id
      - latest_data
      - latitude
      - longitude
      - provider
      - status
      - updated_at
      type: object
    SensorData:
      example:
        updated_at: updated_at
        temperature: 4.145608029883936
        created_at: created_at
        humidity: 2.027123023002322
        watermarks:
        - centibar: 7
          depth: 1
          resistance: 1
        - centibar: 7
          depth: 1
          resistance: 1
        battery: 3.616076749251911
      properties:
        battery:
          type: number
        created_at:
          type: string
        humidity:
          type: number
        temperature:
          type: number
        updated_at:
          type: string
        watermarks:
          items:
            $ref: '#/components/schemas/WatermarkResponse'
          type: array
      required:
      - battery
      - created_at
      - humidity
      - temperature
      - updated_at
      - watermarks
      type: object
    SensorList:
      example:
        pagination:
          next_page: 9
          total_records: 4
          total_pages: 2
          prev_page: 3
          current_page: 7
        data:
        - additional_information:
            key: ""
          latest_data:
            updated_at: updated_at
            temperature: 4.145608029883936
            created_at: created_at
            humidity: 2.027123023002322
            watermarks:
            - centibar: 7
              depth: 1
              resistance: 1
            - centibar: 7
              depth: 1
              resistance: 1
            battery: 3.616076749251911
          updated_at: updated_at
          provider: provider
          latitude: 1.4894159098541704
          created_at: created_at
          id: id
          longitude: 6.84685269835264
          status: online
        - additional_information:
            key: ""
          latest_data:
            updated_at: updated_at
            temperature: 4.145608029883936
            created_at: created_at
            humidity: 2.027123023002322
            watermarks:
            - centibar: 7
              depth: 1
              resistance: 1
            - centibar: 7
              depth: 1
              resistance: 1
            battery: 3.616076749251911
          updated_at: updated_at
          provider: provider
          latitude: 1.4894159098541704
          created_at: created_at
          id: id
          longitude: 6.84685269835264
          status: online
      properties:
        data:
          items:
            $ref: '#/components/schemas/Sensor'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
      - data
      type: object
    SensorStatus:
      enum:
      - online
      - offline
      - unknown
      type: string
      x-enum-varnames:
      - SensorStatusOnline
      - SensorStatusOffline
      - SensorStatusUnknown
...
    WatermarkResponse:
      example:
        centibar: 7
        depth: 1
        resistance: 1
      properties:
        centibar:
          type: integer
        depth:
          type: integer
        resistance:
          type: integer
      required:
      - centibar
      - depth
      - resistance
      type: object
...
  securitySchemes:
    Keycloak:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/auth
          scopes: {}
          tokenUrl: https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/token
      type: oauth2
...